backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
motor
dnspython
python-multipart
passlib[bcrypt]
python-jose[cryptography]
email-validator
